fix: HTTP 413 should not stop event sender (#43)
HTTP status code 413 is considered unrecoverable by our standard status
checking logic. And this is correct -- any individual request that
receives a 413 should be considered invalid.

However, an individual request in the event sender does not necessarily
mean we should shut down the entire sender. In fact, if it is a 413, we
should continue processing the event stream for any subsequent event
batches.