feat: Add ability to force sampling (#58)
Like other SDKS, the PHP SDK can honor sampling ratios to limit how
frequently it emits events. However, since PHP doesn't summarize events,
we need to rely on the relay to handle this.

During that process, the relay receives the event from PHP, deserializes
it into a compatible format, and then sends it through the event
processing architecture.

If nothing else was changed, this would cause the event to be sampled
twice. To prevent this, we are adding a new `ForceSampling` option to
each event which instructs the processor to ignore any provided sampling
ratio for the purposes of sampling, but to continue reporting the
original sampling ratio when encoded.