fix: Fix potential deadlock during processor shutdown (#29)
We try to ensure an orderly shutdown of the processor and all of its
related go routines.

The processor spins up 5 go routines for flushing events. Those workers
post events and then send the results back on a results channel.
Since the only listener of that channel is the same one that processes
shut down requests, you enter a race condition where:

- The dispatcher is waiting for all of the workers to shut down before
  it continues, and
- 1 or more workers are trying to send their final event results to a
  channel no one is listening to.

To resolve this, we provide a cancellable context to the flush workers,
allowing us to inform them when they should abandon their work, and skip
sending results.

This commit also fixes a slight usage issue.

The `FlushBlocking` method suggests that it will block until the
requested flush has completed. However, in practice, it is blocking
until all flush workers have completed and no future pending flushes are
waiting. If flush requests are coming in fast enough, this could
theoretically never complete even though the request had been honored.

To remedy this behavior, we are updating the flush payload event with
the channel that should be closed on completion. This allows for
monitoring of a specific flush request as advertised.